add axes and second camera for it (#23)
fixes #5

I use two camera entities to do it. The main camera is look at the structure. The gizmo camera is look at the coordinates and rendering to the left-bottom corner. This technology is similar to the mini-map in the rpg game.

The axes idea is same to Blender and UR engine to have a axes marker showing the orientation.